Make the config command a StrategyCommand
