+++ guet/commands/config/factory.py
@@ -1,8 +1,9 @@
 from typing import List
 
+from guet.commands.config.set_config_strategy import SetConfigStrategy
+from guet.commands.strategy_command import StrategyCommand
 from guet.settings.settings import Settings
 from guet.commands.command_factory import CommandFactoryMethod
-from guet.commands.config.command import ConfigSetCommand
 
 CONFIG_HELP_MESSAGE = 'usage: guet config [--<key>=<value> ...]'
 
@@ -12,4 +13,4 @@ class ConfigCommandFactory(CommandFactoryMethod):
         return 'Change setting values'
 
     def build(self, args: List[str], settings: Settings):
-        return ConfigSetCommand(args, settings)
+        return StrategyCommand(SetConfigStrategy(args))
